Add files via upload
This is the last assignment of the sir naeem which is to maintain the attendance of the students. it also asks date of the day about which to maintain attendance. it is a filing based -i-e every record in the file about the students and also about the attendance. attendance summary is to tell how much student is present and absent at a particular date.