1st Sem Final CP Lab Project
This is a final project of CP Lab submitted to sir Aleem. This project is basically shopping store. it also contains the logins credentials of the user and the admin where admin portal is change and user portal is change. admin can have access where user does not have. it is a filing based. receipt, inventory, cart ... every imp thing is stored in the files.